+++ RirePE/ControlList.h
@@ -2,34 +2,20 @@
 #define __CONTROL_H__
 
 enum SubControl {
+	STATIC_INFO = 100,
+	BUTTON_OPEN_FORMATVIEW,
+	BUTTON_OPEN_FORMATVIEW2,
 	LISTVIEW_LOGGER,
-	EDIT_EXTRA,
 	BUTTON_CLEAR,
-	BUTTON_EXPORT,
-	EDIT_SEND_HEADER,
-	BUTTON_SEND_IGNORE,
-	BUTTON_SEND_IGNORE_DELETE,
-	EDIT_RECV_HEADER,
-	BUTTON_RECV_IGNORE,
-	BUTTON_RECV_IGNORE_DELETE,
-	LISTVIEW_HEADERS,
-	BUTTON_SEND_BLOCK,
-	BUTTON_SEND_BLOCK_DELETE,
-	BUTTON_RECV_BLOCK,
-	BUTTON_RECV_BLOCK_DELETE,
 	CHECK_SEND,
 	CHECK_RECV,
-	BUTTON_SAVE_FILTER,
 	EDIT_PACKET_SEND,
 	BUTTON_SEND,
 	EDIT_PACKET_RECV,
 	BUTTON_RECV,
+	CHECK_LOCK,
 	BUTTON_INC_SEND,
 	BUTTON_INC_RECV,
-	CHECK_LOCK,
-	CHECK_SHOW_RETURN,
-	CHECK_SHOW_RAWDATA,
-	CHECK_SHOW_INTDATA,
 };
 
 enum ListViewIndex {
