+++ RirePE64/RirePE64.vcxproj
@@ -20,12 +20,14 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\RirePE\ControlList.h" />
+    <ClInclude Include="..\RirePE\FormatGUI.h" />
     <ClInclude Include="..\RirePE\MainGUI.h" />
     <ClInclude Include="..\RirePE\PacketLogger.h" />
     <ClInclude Include="..\RirePE\PacketSender.h" />
     <ClInclude Include="..\RirePE\RirePE.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="..\RirePE\FormatGUI.cpp" />
     <ClCompile Include="..\RirePE\MainGUI.cpp" />
     <ClCompile Include="..\RirePE\PacketLogger.cpp" />
     <ClCompile Include="..\RirePE\PacketSender.cpp" />
